still more modernization, bug fixes, and code style changes.

+++ win/nh3d/NH3DMapView.h
@@ -81,7 +81,7 @@
 	//NSShadow *fontShadow;
 	//NSMutableDictionary *fontAttributes;
 
-	float	cursOpacity;
+	CGFloat	cursOpacity;
 		
 	NH3DMapItem *mapItemValue[MAPVIEWSIZE_COLUMN][MAPVIEWSIZE_ROW];
